app/requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
coverage==7.2.7
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj3-cloudinary-storage==0.0.6
Django==4.2.10
django-allauth==0.54.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.6.0